Revert "added merkle tree"
This reverts commit 152d717ac9cea168126414a7ff66a3eca63badce.
