+++ bazooka/types.go
@@ -1,10 +1,6 @@
 package bazooka
 
-import (
-	"math/big"
-
-	"github.com/BOPR/core"
-)
+import "math/big"
 
 type TypesUserState struct {
 	PubkeyID *big.Int
@@ -16,11 +12,4 @@ type TypesUserState struct {
 type TypesStateMerkleProof struct {
 	State   TypesUserState
 	Witness [][32]byte
-}
-
-// TypesCommitmentInclusionProof is an auto generated low-level Go binding around an user-defined struct.
-type TypesCommitmentInclusionProof struct {
-	Commitment core.CommitmentData
-	Path       *big.Int
-	Witness    [][32]byte
 }
