+++ db/batch.go
@@ -1,12 +1,6 @@
 package db
 
-import (
-	"fmt"
-	"math/big"
-
-	"github.com/BOPR/bazooka"
-	"github.com/BOPR/core"
-)
+import "github.com/BOPR/core"
 
 func (db *DB) GetLatestBatch() (batch core.Batch, err error) {
 	if err := db.Instance.Order("batch_id desc").First(&batch).Error; err != nil {
@@ -64,46 +58,4 @@ func (db *DB) addCommitmentsWithBatchID(commitments []core.Commitment, batchID u
 	return db.Instance.Create(&dbCommitments).Error
 }
 
-// GetCommitmentsForBatch getter for all commitments sent in a batch in order of offset
-func (db *DB) GetCommitmentsForBatch(id uint64) (commitments []core.Commitment, err error) {
-	if err = db.Instance.Model(&commitments).Where("batch_id = ?", id).Order("offset asc").Find(&commitments).Error; err != nil {
-		return
-	}
-	return
-}
-
-func (db *DB) GetLastCommitmentMP(id uint64) (err error) {
-	commitments, err := db.GetCommitmentsForBatch(id)
-	if err != nil {
-		return err
-	}
-
-	var leaves []core.ByteArray
-	for _, commitment := range commitments {
-		leaf, err := commitment.CommitmentData.Leaf()
-		if err != nil {
-			return err
-		}
-		leaves = append(leaves, leaf)
-	}
-
-	tree, err := core.NewTree(leaves)
-	if err != nil {
-		return err
-	}
-
-	lastCommitment := len(commitments) - 1
-	leaf, witness, err := tree.GetWitnessForLeaf(uint64(lastCommitment))
-	if err != nil {
-		return err
-	}
-
-	commitmentMP := bazooka.TypesCommitmentInclusionProof{
-		Commitment: commitments[lastCommitment].CommitmentData,
-		Path:       big.NewInt(int64(lastCommitment)),
-		Witness:    witness,
-	}
-
-	fmt.Println("leaf and witness", leaf, witness)
-	return nil
-}
+// getter for all commitments sent in a batch in order of offset
